Remove network from docker, change @AllArgsConstructor to @RequiredArgsConstructor, change double to BigDecimal

+++ src/main/java/com/example/Bar/converter/DiscountCardConverter.java
@@ -11,7 +11,7 @@ public class DiscountCardConverter {
         final DiscountCardDTO discountCardDTO = new DiscountCardDTO();
         discountCardDTO.setName(userDiscountCardEntity.getName());
         discountCardDTO.setEmail(userDiscountCardEntity.getUserEntity().getEmail());
-        discountCardDTO.setAllSpentMoney(userDiscountCardEntity.getAllSpentMoney());
+        discountCardDTO.setAllSpentMoney(userDiscountCardEntity.getAllSpentMoney().doubleValue());
         discountCardDTO.setDiscount(userDiscountCardEntity.getClientDiscount());
 
         return discountCardDTO;
